--- MyEMS/README.md
@@ -1,13 +1,0 @@
-# MyEMS
-
-This is my current project that I am working on in my free time. My intentions with this project are to teach myself and better understand OOP principles and to prepare me for another project that I plan on undertaking in
-the near future as part of research club at University.
-
-I plan to make a system that would replace something like an excel spreadsheet, as well as something that is capable of filling out payslips each month as well as logging hours. Potentially I would do this with admin/manager accounts
-and then staff accounts so that the staff can log their hours themself but I am yet to decide. I look forward to putting more on here and completing items on the checklist below of features. *? In the to do list indicate features that I may add in the future, not in the initial building.
-
-- [ ] - Add and remove employees + info
-- [ ] - Calculate pay each month/biweekly/weekly?
-- [ ] - Generate Pay-Slips?
-- [ ] - Ability to log hours
-- [ ] - Generate Rotas
