+++ MyEMS/src/DB_Init/User.java
@@ -1,0 +1,31 @@
+package DB_Init;
+
+import java.util.Locale;
+
+public class User {
+
+    private final int id;
+    private final String email, password;
+
+
+    public User(int id, String email, String password){
+        this.id = id;
+        this.email = email;
+        this.password = password;
+    }
+
+
+    public int getId() {
+        return id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+
+}
