+++ MyEMS/src/DB_Init/emsDB.java
@@ -1,0 +1,16 @@
+package DB_Init;
+//import io.github.cdimascio.dotenv.Dotenv;
+//import com.mysql.cj.x.protobuf.MysqlxPrepare;
+import java.math.BigDecimal;
+import java.sql.*;
+import java.util.ArrayList;
+
+public class emsDB {
+
+    private static final String db_url = "jdbc:mysql://127.0.0.1:3306/?user=root";
+    private static final String db_username = "root";
+    private static final String db_password = System.getenv("DB_PASS");
+
+    //public static
+
+}
