--- MyEMS/toDo.md
@@ -1,12 +1,0 @@
- # To Do List
- 
- - [X] Build a file that can test the functions of the employee class 
- - [X] Continue to add functions to the employee class 
- - [X] Work on the front screen for the app
- - [X] Add log in page
- - [ ] Change the fonts and sized of text on log in and registration
- - [ ] Complete Employee class (TBC)
- - [X] Build GUI file
- - [X] Complete main class
- - [ ] Rethink current design - consider more how things will be calculated and stored
- - [ ] Finish building the add employees page
